Производит разбиение на 2 модуля
